+++ changelog.md
@@ -1,6 +1,10 @@
-## 1.1.2
+### 1.1.3
+fix a leader missing problem when the client reconnects
+fix kick and promote buttons can't work properly.
+
+### 1.1.2
 revert accidentally upload content
 fix MNS compat
 
-## 1.1.1
+### 1.1.1
 fix depth issue